--- app/src/main/assets/num.chords
@@ -1,22 +1,0 @@
-10000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00000-00000 key 0
-00000-10000:00000-00000 00000-00000:00000-00000 00000-00000:00000-00000 
-00000-00000:00010-00000 00000-00000:00100-00000 00000-00000:00000-00000 shift SEMICOLON
-00000-00000:00000-10000 00000-00000:00000-00000 00000-00000:00000-00000 key 5
-00000-00000:00000-00000 10000-00000:00000-00000 00000-00000:00000-00000 key 1
-00000-00000:00000-00000 00000-10000:00000-00000 00000-00000:00000-00000 key 2
-00000-00000:00000-00000 00000-00000:10000-00000 00000-00000:00000-00000 key 6
-00000-00000:00000-00000 00000-00000:00100-00000 00000-00000:00000-00000 key PERIOD
-00000-00000:00000-00000 00000-00000:00002-02000 00000-00000:00000-00000 delete line
-00000-00000:00000-00000 00000-00000:00001-00000 00000-00000:00000-00000 key SPACE
-00000-00000:00000-00000 00000-00000:00000-10000 00000-00000:00000-00000 key 7
-00000-00000:00000-00000 00000-00000:00000-01000 00000-00000:00000-00000 key DEL
-00000-00000:00000-00000 00000-00000:00000-00100 00000-00000:00000-00000 key FORWARD_DEL
-00000-00000:00000-00000 00000-00000:00000-00010 00000-00000:00000-00000 key DPAD_DOWN
-00000-00000:00000-00000 00000-00000:00000-00000 10000-00000:00000-00000 key 3
-00000-00000:00000-00000 00000-00000:00000-00000 00000-10000:00000-00000 key 4
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:10000-00000 key 8
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00100-00000 key COMMA
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00001-00000 key DPAD_RIGHT
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00000-10000 key 9
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00000-01000 key DPAD_LEFT
-00000-00000:00000-00000 00000-00000:00000-00000 00000-00000:00000-00100 key DPAD_UP
