+++ app/src/main/java/com/adellica/thumbkeyboard/tsm/Reader.java
@@ -42,71 +42,6 @@ public class Reader {
     public static String writeString(String s) {
         return s;
     }
-        /*
-        ByteArrayOutputStream bao = new ByteArrayOutputStream();
-        ByteArrayInputStream bai = new ByteArrayInputStream(s.getBytes());
-        int c;
-        try {
-            while ((c = bai.read()) >= 0) {
-                switch (c) {
-                    case 0x07:
-                        bao.write('\\');
-                        c = 'a';
-                        break;
-                    case 0x08:
-                        bao.write('\\');
-                        c = 'b';
-                        break;
-                    case 0x0C:
-                        bao.write('\\');
-                        c = 'f';
-                        break;
-                    case 0x0A:
-                        bao.write('\\');
-                        c = 'n';
-                        break;
-                    case 0x0D:
-                        bao.write('\\');
-                        c = 'r';
-                        break;
-                    case 0x09:
-                        bao.write('\\');
-                        c = 't';
-                        break;
-                    case 0x0B:
-                        bao.write('\\');
-                        c = 'v';
-                        break;
-                    case 0x5C:
-                        bao.write('\\');
-                        c = '\\';
-                        break;
-                    case 0x27:
-                        bao.write('\\');
-                        c = '\'';
-                        break;
-                    case 0x22:
-                        bao.write('\\');
-                        c = '"';
-                        break;
-                    case 0x3F:
-                        bao.write('\\');
-                        c = '?';
-                        break;
-                    default:
-                }
-                if (isPrintableChar(c)) {
-                    bao.write(c);
-                } else {
-                    bao.write(String.format("\\x%02x", c).getBytes());
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return bao.toString();
-    }
-    */
 
     final InputStream is;
